moved Choices class to sdk.queries.__init__.py
